271---Avoiding-Component-Function-Executions-with-Clever-Structuring
