+++ 13---A-Look-Behind-The-Scenes-Of-React---And---Optimization-Techniques/01-starting-project/src/App.jsx
@@ -3,31 +3,23 @@ import { useState } from 'react';
 import Counter from './components/Counter/Counter.jsx';
 import Header from './components/Header.jsx';
 import { log } from './log.js';
+import ConfigureCounter from './components/Counter/ConfigureCounter.jsx';
 
 function App() {
   log('<App /> rendered');
 
-  const [enteredNumber, setEnteredNumber] = useState(0);
   const [chosenCount, setChosenCount] = useState(0);
 
-  function handleChange(event) {
-    setEnteredNumber(+event.target.value);
-  }
 
-  function handleSetClick() {
-    setChosenCount(enteredNumber);
-    setEnteredNumber(0);
+  function handleSetCount(newCount) {
+    setChosenCount(newCount);
   }
 
   return (
     <>
       <Header />
       <main>
-        <section id="configure-counter">
-          <h2>Set Counter</h2>
-          <input type="number" onChange={handleChange} value={enteredNumber} />
-          <button onClick={handleSetClick}>Set</button>
-        </section>
+        <ConfigureCounter onSet={handleSetCount} />
         <Counter initialCount={chosenCount} />
       </main>
     </>
