+++ 13---A-Look-Behind-The-Scenes-Of-React---And---Optimization-Techniques/01-starting-project/src/components/Counter/Counter.jsx
@@ -27,7 +27,7 @@ function isPrime(number) {
   return true;
 }
 
-const Counter = memo(function Counter({ initialCount }) {
+const Counter = (function Counter({ initialCount }) {
   log('<Counter /> rendered', 1);
   const initialCountIsPrime = isPrime(initialCount);
